Move OCP hack in its own test file
Keeping downstream only change to a separated file avoid conflicts
when merging upstream changes.
